requirements.txt
tkinter
nltk
textblob
newspaper3k
lxml_html_clean